Refactor line style.
