+++ src/styles/line.ts
@@ -3,6 +3,7 @@ import { Stroke, Style } from 'ol/style'
 import { getTrambusLineNumber, lineColors, lineDimmedColors } from './common'
 import type { FeatureLike } from 'ol/Feature'
 import { SelectedTrambusLine } from '@/model/selected-line.model'
+import type { TravelTimeModel } from '@/model/travel-time.model'
 
 export type LineState = 'selected' | 'normal' | 'unselected' | 'hidden'
 
@@ -71,6 +72,24 @@ export function trambusLineViewStyleFunction(
     lineState = 'selected'
   } else {
     lineState = displayOtherLine ? 'unselected' : 'hidden'
+  }
+  return trambusLineStyle(lineNumber, lineState, is3D)
+}
+
+export function trambusLineTravelTimesViewStyleFunction(
+  feature: FeatureLike,
+  selectedTravelTime: TravelTimeModel,
+  is3D: boolean
+): Style[] {
+  const lineNumber = getTrambusLineNumber(feature) as LineNumber
+  let lineState: LineState = 'normal'
+
+  if (selectedTravelTime == null) {
+    lineState = 'normal'
+  } else if (getTrambusLineNumber(feature) == selectedTravelTime.line) {
+    lineState = 'selected'
+  } else {
+    lineState = 'unselected'
   }
   return trambusLineStyle(lineNumber, lineState, is3D)
 }
